@@ -5,40 +5,46 @@
     "ionic:build": "ionic-app-scripts build",
     "ionic:serve": "ionic-app-scripts serve"
   },
-  "config": {
-  },
+  "config": {},
   "dependencies": {
-    "@angular/compiler": "2.4.8",
-    "@angular/compiler-cli": "2.4.8",
-    "@angular/common": "2.4.8",
-    "@angular/core": "2.4.8",
-    "@angular/forms": "2.4.8",
-    "@angular/http": "2.4.8",
-    "@angular/platform-browser": "2.4.8",
-    "@angular/platform-browser-dynamic": "2.4.8",
-    "@angular/platform-server": "2.4.8",
-    "@angular/router": "3.4.8",
-    "@ionic/storage": "1.1.6",
-    "angular2-jwt": "0.1.28",
-    "angularfire2": "2.0.0-beta.7",
+    "@angular/animations": "4.0.2",
+    "@angular/common": "4.0.2",
+    "@angular/compiler": "4.0.2",
+    "@angular/compiler-cli": "4.0.2",
+    "@angular/core": "4.0.2",
+    "@angular/forms": "4.0.2",
+    "@angular/http": "4.0.2",
+    "@angular/platform-browser": "4.0.2",
+    "@angular/platform-browser-dynamic": "4.0.2",
+    "@angular/platform-server": "4.0.2",
+    "@angular/router": "4.0.2",
+    "@ionic-native/core": "^3.6.1",
+    "@ionic-native/file": "^3.6.1",
+    "@ionic-native/native-storage": "^3.6.1",
+    "@ionic-native/network": "3.6.1",
+    "@ionic-native/push": "^3.6.1",
+    "@ionic-native/splash-screen": "3.6.1",
+    "@ionic-native/status-bar": "3.6.1",
+    "@ionic-native/transfer": "^3.6.1",
+    "@ionic/storage": "2.0.1",
+    "angular2-jwt": "0.2.3",
+    "angularfire2": "4.0.0-rc.0",
     "es6-shim": "0.35.1",
-    "firebase": "3.5.2",
-    "ionic-angular": "2.2.0",
-    "ionic-native": "2.8.0",
+    "firebase": "3.6.6",
+    "ionic-angular": "3.1.1",
     "ionicons": "3.0.0",
     "log4javascript": "1.4.15",
     "moment": "2.15.1",
     "reflect-metadata": "0.1.8",
-    "rxjs": "5.0.1",
+    "rxjs": "5.1.1",
     "socket.io-client": "^1.7.2",
+    "sw-toolbox": "3.4.0",
     "underscore": "1.8.3",
-    "zone.js": "0.7.2",
-    "sw-toolbox": "3.4.0"
+    "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "1.1.4",
-    "typescript": "2.0.9",
-    "@types/lodash": "4.14.37",
+    "@ionic/app-scripts": "1.3.4",
+    "@types/lodash": "^4.14.64",
     "@types/request": "0.0.30",
     "browserify": "13.1.0",
     "browserify-istanbul": "2.0.0",
@@ -61,6 +67,7 @@
     "ts-node": "1.3.0",
     "tsify": "1.0.3",
     "tslint": "3.14.0",
+    "typescript": "~2.2.1",
     "watchify": "3.7.0",
     "xml2js": "0.4.17"
   },
