@@ -18,14 +18,14 @@
 * // App Root 
 * import { AppComponent }   from './app.component';
 * // Feature Modules
-* import { SDKModule }      from './shared/sdk/sdk.module';
+* import { SDK[Browser|Node|Native]Module } from './shared/sdk/sdk.module';
 * // Import Routing
 * import { routing }        from './app.routing';
 * @NgModule({
 *  imports: [
 *    BrowserModule,
 *    routing,
-*    SDKModule.forRoot()
+*    SDK[Browser|Node|Native]Module.forRoot()
 *  ],
 *  declarations: [ AppComponent ],
 *  bootstrap:    [ AppComponent ]
@@ -38,11 +38,12 @@ import { ErrorHandler } from './services/core/error.service';
 import { LoopBackAuth } from './services/core/auth.service';
 import { LoggerService } from './services/custom/logger.service';
 import { SDKModels } from './services/custom/SDKModels';
-import { InternalStorage } from './storage/internal.storage';
+import { InternalStorage, SDKStorage } from './storage/storage.swaps';
 import { HttpModule } from '@angular/http';
 import { CommonModule } from '@angular/common';
 import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CookieBrowser } from './storage/cookie.browser';
+import { StorageBrowser } from './storage/storage.browser';
 import { HsyListingApi } from './services/custom/HsyListing';
 import { HsyUserApi } from './services/custom/HsyUser';
 /**
@@ -63,7 +64,10 @@ import { HsyUserApi } from './services/custom/HsyUser';
   ]
 })
 export class SDKBrowserModule {
-  static forRoot(): ModuleWithProviders {
+  static forRoot(internalStorageProvider: any = {
+    provide: InternalStorage,
+    useClass: CookieBrowser
+  }): ModuleWithProviders {
     return {
       ngModule  : SDKBrowserModule,
       providers : [
@@ -73,7 +77,8 @@ export class SDKBrowserModule {
         SDKModels,
         HsyListingApi,
         HsyUserApi,
-        { provide: InternalStorage, useClass: CookieBrowser }
+        internalStorageProvider,
+        { provide: SDKStorage, useClass: StorageBrowser }
       ]
     };
   }
@@ -85,4 +90,6 @@ export class SDKBrowserModule {
 export * from './models/index';
 export * from './services/index';
 export * from './lb.config';
-
+export * from './storage/storage.swaps';
+export { CookieBrowser } from './storage/cookie.browser';
+export { StorageBrowser } from './storage/storage.browser';
