@@ -2,23 +2,23 @@
 
 declare var Object: any;
 export interface HsyUserInterface {
-  avatarId?: string;
-  id: string;
-  name?: string;
-  pushNotificationRegIds?: Array<string>;
-  weixin?: string;
-  created?: Date;
-  lastUpdated?: Date;
+  "avatarId"?: string;
+  "id": string;
+  "name"?: string;
+  "pushNotificationRegIds"?: Array<any>;
+  "weixin"?: string;
+  "created"?: Date;
+  "lastUpdated"?: Date;
 }
 
 export class HsyUser implements HsyUserInterface {
-  avatarId: string = '';
-  id: string = '';
-  name: string = '';
-  pushNotificationRegIds: Array<string> = <any>[];
-  weixin: string = '';
-  created: Date = new Date(0);
-  lastUpdated: Date = new Date(0);
+  "avatarId": string = '';
+  "id": string = '';
+  "name": string = '';
+  "pushNotificationRegIds": Array<any> = <any>[];
+  "weixin": string = '';
+  "created": Date = new Date(0);
+  "lastUpdated": Date = new Date(0);
   constructor(data?: HsyUserInterface) {
     Object.assign(this, data);
   }
@@ -37,7 +37,7 @@ export class HsyUser implements HsyUserInterface {
   **/
   public static factory(data: HsyUserInterface): HsyUser{
     return new HsyUser(data);
-  }  
+  }
   /**
   * @method getModelDefinition
   * @author Julien Ledun
@@ -50,31 +50,31 @@ export class HsyUser implements HsyUserInterface {
       name: 'HsyUser',
       plural: 'HsyUsers',
       properties: {
-        avatarId: {
+        "avatarId": {
           name: 'avatarId',
           type: 'string'
         },
-        id: {
+        "id": {
           name: 'id',
           type: 'string'
         },
-        name: {
+        "name": {
           name: 'name',
           type: 'string'
         },
-        pushNotificationRegIds: {
+        "pushNotificationRegIds": {
           name: 'pushNotificationRegIds',
-          type: 'Array&lt;string&gt;'
+          type: 'Array&lt;any&gt;'
         },
-        weixin: {
+        "weixin": {
           name: 'weixin',
           type: 'string'
         },
-        created: {
+        "created": {
           name: 'created',
           type: 'Date'
         },
-        lastUpdated: {
+        "lastUpdated": {
           name: 'lastUpdated',
           type: 'Date'
         },
