@@ -3,10 +3,12 @@ import { Injectable } from '@angular/core';
 import { HsyListing } from '../../models/HsyListing';
 import { HsyUser } from '../../models/HsyUser';
 
+export interface Models { [name: string]: any }
+
 @Injectable()
 export class SDKModels {
 
-  private models: { [name: string]: any } = {
+  private models: Models = {
     HsyListing: HsyListing,
     HsyUser: HsyUser,
     
@@ -14,5 +16,13 @@ export class SDKModels {
 
   public get(modelName: string): any {
     return this.models[modelName];
+  }
+
+  public getAll(): Models {
+    return this.models;
+  }
+
+  public getModelNames(): string[] {
+    return Object.keys(this.models);
   }
 }
