@@ -1,14 +1,26 @@
 import { Injectable } from '@angular/core';
+export interface CookieInterface { [key: string]: any }
 /**
+* @author Jonathan Casarrubias <twitter:@johncasarrubias> <github:@mean-expert-official>
 * @module CookieBrowser
-* @author Jonathan Casarrubias
 * @license MIT
 * @description
-* Stand-alone cookie service for browsers
+* This module handle cookies, it will be provided using DI Swapping according the
+* SDK Socket Driver Available currently supporting Angular 2 for web and NativeScript 2.
 **/
 @Injectable()
 export class CookieBrowser {
-  private cookies: { [key: string]: any } = {};
+  /**
+   * @type {CookieInterface}
+   **/
+  private cookies: CookieInterface = {};
+  /**
+   * @method get
+   * @param {string} key Cookie key name
+   * @return {any}
+   * @description
+   * The getter will return any type of data persisted in cookies.
+   **/
   get(key: string): any {
     if (!this.cookies[key]) {
       let cookie = window.document
@@ -18,22 +30,43 @@ export class CookieBrowser {
         return null;
       }
 
-      this.cookies[key] = this.parse(cookie.split('=').pop());
+      this.cookies[key] = this.parse(cookie.split('=').slice(1).join('='));
     }
 
     return this.cookies[key];
   }
-
-  set(key: string, value: any, expires?: Date) {
+  /**
+   * @method set
+   * @param {string} key Cookie key name
+   * @param {any} value Any value
+   * @param {Date=} expires The date of expiration (Optional)
+   * @return {void}
+   * @description
+   * The setter will return any type of data persisted in cookies.
+   **/
+  set(key: string, value: any, expires?: Date): void {
     this.cookies[key] = value;
-    let cookie = `${key}=${value}${expires ? `; expires=${ expires.toUTCString() }` : ''}`;
+    let cookie = `${key}=${value}; path=/${expires ? `; expires=${ expires.toUTCString() }` : ''}`;
     window.document.cookie = cookie;
   }
-
+  /**
+   * @method remove
+   * @param {string} key Cookie key name
+   * @return {void}
+   * @description
+   * This method will remove a cookie from the client.
+   **/
   remove(key: string) {
-    document.cookie = key + '=; expires=Thu, 01 Jan 1970 00:00:01 GMT;';
+    document.cookie = key + '=; path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT;';
   }
-
+  /**
+   * @method parse
+   * @param {any} value Input data expected to be JSON
+   * @return {void}
+   * @description
+   * This method will parse the string as JSON if possible, otherwise will
+   * return the value itself.
+   **/
   private parse(value: any) {
     try {
         return JSON.parse(value);
