@@ -3,7 +3,7 @@ import {NavController, Modal, ModalController, Platform} from "ionic-angular";
 import {ChatsTabPage} from "../chats-tab/chats-tab.page";
 import {ListingsTabPage} from "../listings-tab/listings-tab.page";
 import {SettingsTabPage} from "../settings-tab/settings-tab.page";
-import {Network} from "ionic-native";
+import {Network} from "@ionic-native/network";
 import {DisconnectModal} from "./disconnect.modal";
 import {Subscription} from "rxjs";
 import {QrCodeTabPage} from "../qrcode-tab/qrcode-tab-page";
@@ -30,7 +30,8 @@ export class TabsPage implements OnInit, OnDestroy {
   constructor(private nav:NavController,
               private modalCtrl: ModalController,
               private platform:Platform,
-              private navController: NavController) {
+              private navController: NavController,
+              private network: Network) {
   }
 
   ngOnInit():void {
@@ -42,11 +43,11 @@ export class TabsPage implements OnInit, OnDestroy {
       if (this.platform.is("ios") || this.platform.is("android")) {
         this.disconnectModal = this.modalCtrl.create(DisconnectModal);
 
-        this.onDisconnect = Network.onDisconnect().subscribe(() => {
+        this.onDisconnect = this.network.onDisconnect().subscribe(() => {
           this.disconnectModal.present();
         });
 
-        this.onConnect = Network.onConnect().subscribe(() => {
+        this.onConnect = this.network.onConnect().subscribe(() => {
           this.disconnectModal.dismiss();
         });
       }
