@@ -1,7 +1,7 @@
 import {Injectable} from "@angular/core";
 import {Message} from "../../models/models";
 import {Observable} from "rxjs/Observable";
-import {AngularFire} from "angularfire2/angularfire2";
+import {AngularFireDatabase} from "angularfire2/database";
 
 @Injectable()
 export class IMessageService {
@@ -20,12 +20,12 @@ export class IMessageService {
 
 @Injectable()
 export class FirebaseMessageService implements IMessageService {
-  constructor(private af:AngularFire) {
+  constructor(private afdb:AngularFireDatabase) {
   }
 
   // TODO(xinbenlv): optimize for performance
   observableMessagesByThreadId(threadId:string):Observable<Message[]> {
-    return this.af.database.list("/messages", {
+    return this.afdb.list("/messages", {
       query: {
         orderByChild: 'sentAt'
       }
@@ -39,7 +39,7 @@ export class FirebaseMessageService implements IMessageService {
 
   // TODO(xinbenlv): optimize for performance
   observableBadgeCounter(threadId:string, lastCheckTime:number):Observable<number> {
-    return this.af.database.list("/messages", {
+    return this.afdb.list("/messages", {
       query: {
         orderByChild: 'sentAt',
         // TODO(xinbenlv): startAt with a key in AngularFire2
@@ -55,7 +55,7 @@ export class FirebaseMessageService implements IMessageService {
 
   createMessage(message:Message):Promise<void> {
     if (message.id) {
-      return this.af.database.object("/messages/" + message.id).update(message) as Promise<void>;
+      return this.afdb.object("/messages/" + message.id).update(message) as Promise<void>;
     } else {
       // TODO(xinbenlv): handle when message.id is not set
       // currently do nothing
